backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
httpx
python-dotenv
